+++ velo/model/medium.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import
 from datetime import datetime
 
 from mongokit import INDEX_GEO2D
@@ -17,16 +18,18 @@ class Medium(Document):
             'longitude': float,
             'latitude': float,
             },
+        'mime_type': basestring,
         }
 
     gridfs = {
-        'files': ['source']
+        'files': ['content']
         }
 
     required_fields = [
         'creation_datetime',
         'location.longitude',
         'location.latitude',
+        'mime_type'
         ]
 
     default_values = {