+++ velo/tests/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import
 import os
 
 
 def setUpPackage():
     os.environ['MONGO_URI'] = 'mongodb://localhost/velobstacles'
+    os.environ['VELO_DB_NAME'] = 'velobstacles'