+++ velo/tests/functional/test_index.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import
 from velo.tests.functional import TestController