+++ velo/tests/unittests/test_media_view.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
-from bson.binary import Binary
+from __future__ import absolute_import
+import pkg_resources
+
+from cgi import FieldStorage
 
 from pyramid import testing
 from pyramid.httpexceptions import HTTPNotFound
@@ -13,8 +16,10 @@ class TestMedia(TestBase):
         super(TestMedia, self).setUp()
         medium = self.db.Medium()
         medium.location = {'latitude': 45.522706, 'longitude': -73.583885}
+        medium.mime_type = 'image/png'
         medium.save()
-        medium.fs.source = Binary(data='asdf')
+        fp = pkg_resources.resource_stream('velo', 'static/img/trollface.png')
+        medium.fs.put(fp, filename='content')
 
         self.medium_id = medium._id
 
@@ -24,6 +29,7 @@ class TestMedia(TestBase):
         return MediaView(request.context, request)
 
     def tearDown(self):
+        super(TestMedia, self).tearDown()
         self.db.medium.remove({})
         testing.tearDown()
 
@@ -51,6 +57,32 @@ class TestMedia(TestBase):
         self.assertRaises(HTTPNotFound, view.show, '50a85b229978d063d7a3dd04')
 
     def test_show(self):
+        medium = self.db.Medium.find_one({'_id': self.medium_id})
         view = self.get_view()
-        result = view.show(self.medium_id)
+        result = view.show(str(self.medium_id))
         self.assertTrue(result)
+        self.assertEqual(medium.location, result['location'])
+        self.assertEqual(str(medium._id), result['id'])
+        self_link = 'http://example.com/media/%s' % self.medium_id
+        self.assertEqual(
+            {'self': self_link, 'content': self_link + '/content'},
+            result['links'],
+            )
+        self.assertEqual(5, len(result.keys()))
+
+    def test_create(self):
+        view = self.get_view()
+        content = FieldStorage()
+        content.file = pkg_resources.resource_stream(
+            'velo',
+            'static/img/trollface.png'
+            )
+        view.request.POST.update({
+            'longitude': 12.3,
+            'latitude': -45.4,
+            'content': content,
+            })
+        view.create()
+
+    def test_update(self):
+        self.get_view()