+++ CHANGES.rst
@@ -1,3 +1,4 @@
+* Add medium content sub resource which returns medium content;
 * Add medium resource with endpoint: /media;
 * Add medium model with a draft structure;
 * Add model configuration;